+++ src/lib/clib/memory.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "stdint.h"
+#include "stddef.h"
 
-void far* memcpy(void far* dst, const void far* src, uint16_t num);
-void far* memset(void far* ptr, int value, uint16_t num);
-int memcmp(const void far* ptr1, const void far* ptr2, uint16_t num);
+void far* memcpy(void far* dst, const void far* src, size_t num);
+void far* memset(void far* ptr, int value, size_t num);
+int memcmp(const void far* ptr1, const void far* ptr2, size_t num);
